增加BaseDao对金钱的适配，优化金钱的存取方式，将Test文件写为一份

+++ src/com/it/Agile/bean/TB_Dish.java
@@ -1,20 +1,30 @@
 package com.it.Agile.bean;
 
+/**
+ * @author youwannapieceme
+ * @LastModifiedBy IGK_DLS
+ * @LastModifiedDate 2019/04/24
+ */
+import java.math.BigDecimal;
+
 public class TB_Dish {
-	
+
 	private String dishname;
-	private float dishprice;
+	private BigDecimal dishprice;
+
 	public String getDishname() {
 		return dishname;
 	}
+
 	public void setDishname(String dishname) {
 		this.dishname = dishname;
 	}
-	public float getDishprice() {
+
+	public BigDecimal getDishprice() {
 		return dishprice;
 	}
-	public void setDishprice(float dishprice) {
+
+	public void setDishprice(BigDecimal dishprice) {
 		this.dishprice = dishprice;
 	}
-	
 }
